test: split build-phase tests into listen_for_node_build.rs and update add tests to reflect separate add/build steps
